requirements.txt
beautifulsoup4==4.12.2
black==23.3.0
faiss-cpu==1.7.4
flake8==6.0.0
joblib==1.3.0
jsonschema==4.17.3
langchain==0.0.209
mypy==1.4.1
nltk==3.8.1
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pydantic==1.10.9
streamlit==1.24.0
streamlit-chat==0.1.1
tiktoken==0.4.0
toml==0.10.2
tqdm==4.65.0
pinecone-client==2.2.2
python-decouple==3.8
